matrixQTL snakemake dev, moving to new dirs
